feat: Refactor DotProduct function and add unit tests for dot product validation
